eth-flask-demo/requirements.txt
flask==2.0.1
werkzeug==2.0.3
web3==6.0.0
python-dotenv==0.19.0
ipfshttpclient==0.7.0
requests>=2.31.0
eth-account>=0.8.0
eth-utils>=2.1.0
eth-typing>=3.0.0
eth-abi>=4.0.0

